feat(misc: scripts: clean pacaur cache): skip spotify

+++ misc/scripts/clean_pacaur_cache.bash
@@ -1,11 +1,13 @@
 #!/usr/bin/env bash
-set -ue
+set -ueo pipefail
 sudo pacman -Scc
 cd ~/.cache/pacaur/
 echo -n "Cache size before:	"
 du -sh 2>/dev/null || true
 for file in */{*.tar*,pkg,src,*.rpm,*.deb} ; do
-	rm -rf $file
+	if [[ ! $(grep -i -e spotify <<< $file) ]] ; then
+		rm -rf $file
+	fi
 done
 echo -n "Cache size after:	"
 du -sh
